% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lavaanPlot.R
\name{buildCall}
\alias{buildCall}
\title{Builds the Diagrammer function call.}
\usage{
buildCall(
  model = model,
  name = name,
  labels = labels,
  graph_options = graph_options,
  node_options = node_options,
  edge_options = edge_options,
  coefs,
  sig,
  stand,
  covs,
  stars,
  digits,
  conf.int,
  edge_styles,
  ...
)
}
\arguments{
\item{model}{A model fit object of class lavaan.}

\item{name}{A string of the name of the plot.}

\item{labels}{An optional named list of variable labels fit object of class lavaan.}

\item{graph_options}{A named list of graph options for Diagrammer syntax. See for options here https://graphviz.org/docs/graph/}

\item{node_options}{A named list of node options for Diagrammer syntax.}

\item{edge_options}{A named list of edge options for Diagrammer syntax.}

\item{coefs}{whether or not to include significant path coefficient values in diagram}

\item{sig}{significance level for determining what significant paths are}

\item{stand}{Should the coefficients being used be standardized coefficients}

\item{covs}{Should model covariances be included in the diagram}

\item{stars}{a character vector indicating which parameters should include significance stars be included for regression paths, latent paths, or covariances. Include which of the 3 you want ("regress", "latent", "covs"), default is none.}

\item{digits}{A number indicating the desired number of digits for the coefficient values in the plot}

\item{conf.int}{A character vector indicating which parameters should have their CIs depicted in the plot. Available choices: "reg", "regression", "lat", "latent"}

\item{edge_styles}{Whether or not to depict non-significant coefficients with dashed paths}

\item{...}{additional arguments to be passed to \code{buildPaths}}
}
\value{
A string specifying the path diagram for \code{model}
}
\description{
Builds the Diagrammer function call.
}
