% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lavaanPlot.R
\name{buildLabels}
\alias{buildLabels}
\title{Adds variable labels to the Diagrammer plot function call.}
\usage{
buildLabels(label_list)
}
\arguments{
\item{label_list}{A named list of variable labels.}
}
\description{
Adds variable labels to the Diagrammer plot function call.
}
