% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lavaanPlot.R
\name{buildPaths}
\alias{buildPaths}
\title{Extracts the paths from the lavaan model.}
\usage{
buildPaths(fit, coefs = coefs, sig = sig, stand = stand, covs = covs,
  stars = stars)
}
\arguments{
\item{fit}{A model fit object of class lavaan.}

\item{coefs}{whether or not to include significant path coefficient values in diagram}

\item{sig}{significance level for determining what significant paths are}

\item{stand}{Should the coefficients being used be standardized coefficients}

\item{covs}{Should model covariances be included in the diagram}

\item{stars}{Should significance stars be included}
}
\description{
Extracts the paths from the lavaan model.
}

