% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lavaanPlot2.R
\name{convert_graph}
\alias{convert_graph}
\title{Uses the diagrammeR functions to turn the ndf and edf into dot}
\usage{
convert_graph(ndf, edf, graph_options)
}
\arguments{
\item{ndf}{A node data frame created by \code{create_nodes}}

\item{edf}{An edge data frame created by \code{create_edges}}

\item{graph_options}{a named list of graphviz graph attributes}
}
\value{
DOT specification of model
}
\description{
Uses the diagrammeR functions to turn the ndf and edf into dot
}
