% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper.R
\name{create_edge_styles}
\alias{create_edge_styles}
\title{Creates  edge styles}
\usage{
create_edge_styles(pvals, edge_styles = F)
}
\arguments{
\item{pvals}{a vector of p values}

\item{edge_styles}{boolean, sets edge styles based on p-values}
}
\description{
Creates  edge styles
}
