% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lavaanPlot2.R
\name{create_edges}
\alias{create_edges}
\title{Creates edge data frame and adds formatting}
\usage{
create_edges(coefs, ndf, edge_options, coef_labels = FALSE)
}
\arguments{
\item{coefs}{a coefficient table from lavaan model created by \code{extract_coefs}}

\item{ndf}{A node data frame created by \code{create_nodes}}

\item{edge_options}{a named list of graphviz edge attributes}

\item{coef_labels}{whether to label edges with coefficient values}
}
\value{
an edge data frame
}
\description{
Creates edge data frame and adds formatting
}
