% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lavaanPlot2.R
\name{create_grviz}
\alias{create_grviz}
\title{Creates the grViz dot language code for plotting}
\usage{
create_grviz(
  model,
  labels = labels,
  include = include,
  graph_options = graph_options,
  node_options = node_options,
  edge_options = edge_options,
  stand = stand,
  ...
)
}
\arguments{
\item{model}{A model fit object of class lavaan.}

\item{labels}{An optional named list of variable labels.}

\item{include}{which parameters to include in the plot. Default is all regression and latent relationships. "covs" will also include covariances, while "all will also include error variances.}

\item{graph_options}{a named list of graphviz graph attributes}

\item{node_options}{a named list of graphviz node attributes}

\item{edge_options}{a named list of graphviz edge attributes}

\item{stand}{Should the coefficients being used be standardized coefficients}

\item{...}{Additional arguments to be passed to \code{create_edges}}
}
\value{
A string specifying the path diagram for \code{model}
}
\description{
Creates the grViz dot language code for plotting
}
