% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lavaanPlot2.R
\name{create_nodes}
\alias{create_nodes}
\title{Creates node data frame and adds formatting}
\usage{
create_nodes(coefs, labels = NULL, node_options)
}
\arguments{
\item{coefs}{a coefficient table from lavaan model created by \code{extract_coefs}}

\item{labels}{An optional list of labels}

\item{node_options}{a named list of graphviz node attributes, or a data frame of node options created by \code{formatting},}
}
\value{
an edge data frame
}
\description{
Creates node data frame and adds formatting
}
