% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lavaanPlot2.R
\name{extract_coefs}
\alias{extract_coefs}
\title{Creates a data frame of the parameter table from lavaan model}
\usage{
extract_coefs(model, include = NULL, stand = FALSE)
}
\arguments{
\item{model}{A fitted model of class lavaan}

\item{include}{which parameters to include in the plot. Default is all regression and latent relationships. "covs" will also include covariances, while "all" will also include error variances.}

\item{stand}{Should the coefficients being used be standardized coefficients}
}
\value{
a data frame with lavaan model parameters
}
\description{
Creates a data frame of the parameter table from lavaan model
}
