% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lavaanPlot2.R
\name{formatting}
\alias{formatting}
\title{Enables conditional formatting for different parts of the model}
\usage{
formatting(..., type, groups)
}
\arguments{
\item{...}{lists of node or edge options for each of the groups}

\item{type}{type of conditional formatting being used, node, edge, or custom; custom only works with edges}

\item{groups}{character vector of the names of custom groups, with nodes and edges default values are set and you need to match the order: for nodes: c("latent", "obs"), for edges: c("regress", "latent", "covs"). For custom groups of edges, you must match names that you pre-multiply with coefficients in your model specification.}
}
\value{
a formatting data frame that can work with the create_nodes and create_edges functions
}
\description{
Enables conditional formatting for different parts of the model
}
