% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lavaanPlot.R, R/lavaanPlot2.R
\name{sig_stars}
\alias{sig_stars}
\title{Generates standard significance stars}
\usage{
sig_stars(pvals)

sig_stars(pvals)
}
\arguments{
\item{pvals}{a vector of p values}
}
\description{
Generates standard significance stars

Generates standard significance stars
}
