% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lavaanPlot.R
\name{buildCall}
\alias{buildCall}
\title{Builds the Diagrammer function call.}
\usage{
buildCall(name = "plot", model, labels = NULL, graph_options, node_options,
  edge_options)
}
\arguments{
\item{name}{A string of the name of the plot.}

\item{model}{A model fit object of class lavaan.}

\item{labels}{An optionalnamed list of variable labels fit object of class lavaan.}

\item{graph_options}{A named list of graph options for Diagrammer syntax.}

\item{node_options}{A named list of node options for Diagrammer syntax.}

\item{edge_options}{A named list of edge options for Diagrammer syntax..}
}
\value{
A string specifying the path diagram for \code{model}
}
\description{
Builds the Diagrammer function call.
}

