% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lavaanPlot.R
\name{buildCall}
\alias{buildCall}
\title{Builds the Diagrammer function call.}
\usage{
buildCall(
  model = model,
  name = name,
  labels = labels,
  graph_options = list(overlap = "true", fontsize = "10"),
  node_options = list(shape = "box"),
  edge_options = list(color = "black"),
  ...
)
}
\arguments{
\item{model}{A model fit object of class lavaan.}

\item{name}{A string of the name of the plot.}

\item{labels}{An optional named list of variable labels fit object of class lavaan.}

\item{graph_options}{A named list of graph options for Diagrammer syntax.}

\item{node_options}{A named list of node options for Diagrammer syntax.}

\item{edge_options}{A named list of edge options for Diagrammer syntax.}

\item{...}{additional arguments to be passed to \code{buildPaths}}
}
\value{
A string specifying the path diagram for \code{model}
}
\description{
Builds the Diagrammer function call.
}
