% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lavaanPlot.R
\name{buildPaths}
\alias{buildPaths}
\title{Extracts the paths from the lavaan model.}
\usage{
buildPaths(fit)
}
\arguments{
\item{fit}{A model fit object of class lavaan.}
}
\description{
Extracts the paths from the lavaan model.
}

