% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lavaanPlot.R
\name{buildPaths}
\alias{buildPaths}
\title{Extracts the paths from the lavaan model.}
\usage{
buildPaths(fit, coefs, sig, stand, covs, stars, digits, conf.int, edge_styles)
}
\arguments{
\item{fit}{A model fit object of class lavaan.}

\item{coefs}{whether or not to include significant path coefficient values in diagram}

\item{sig}{significance level for determining what significant paths are}

\item{stand}{Should the coefficients being used be standardized coefficients}

\item{covs}{Should model covariances be included in the diagram}

\item{stars}{a character vector indicating which parameters should include significance stars be included for regression paths, latent paths, or covariances. Include which of the 3 you want ("regress", "latent", "covs"), default is none.}

\item{digits}{A number indicating the desired number of digits for the coefficient values in the plot}

\item{conf.int}{whether or not to include confidence intervals of path coefficients}

\item{edge_styles}{Whether or not to depict non-significant coefficients with dashed paths}
}
\description{
Extracts the paths from the lavaan model.
}
