% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotExportFunctions.R
\name{embed_plot_pdf}
\alias{embed_plot_pdf}
\title{Embeds a plot into an rmarkdown pdf}
\usage{
embed_plot_pdf(plot, path)
}
\arguments{
\item{plot}{plot object created by \code{lavaanPlot}}

\item{path}{Filename to save the image}
}
\description{
Embeds a plot into an rmarkdown pdf
}
