% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lavaanPlot.R
\name{lavaanPlot}
\alias{lavaanPlot}
\title{Plots lavaan path model with DiagrammeR}
\usage{
lavaanPlot(name = "plot", model, labels = NULL, ...)
}
\arguments{
\item{name}{A string of the name of the plot.}

\item{model}{A model fit object of class lavaan.}

\item{labels}{An optional named list of variable labels.}

\item{...}{Additional arguments to be called to \code{buildCall} and \code{buildPaths}}
}
\value{
A Diagrammer plot of the path diagram for \code{model}
}
\description{
Plots lavaan path model with DiagrammeR
}
